+++ query_preprocessing/requirements.txt
@@ -43,7 +43,6 @@ langsmith==0.1.3
 marshmallow==3.20.2
 multidict==6.0.5
 mypy-extensions==1.0.0
-numpy==1.26.4
 packaging==23.2
 platformdirs==3.10.0
 passlib==1.7.4
